# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CED OPS API",
    "description": "REST API for the CED (Continuous Electro-Deposition) Operations management system. Configuration-driven report engine over a module hierarchy (module types, modules, template versions, processes, process parameters and global parameters), report sessions and completed reports, plus dashboard, unified search, analytics and role-based access control.",
    "termsOfService": "https://aerotech.com/terms",
    "contact": {
      "name": "AeroTech Support",
      "url": "https://aerotech.com",
      "email": "support@aerotech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://aerotech.com/license"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Default Server URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/processes/{processId}/parameters": {
      "get": {
        "tags": [
          "Process Parameter Master"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProcessParameterResponse"
                      }
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Process Parameter Master"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Roles: SUPER_ADMIN, ADMIN",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProcessParameterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProcessParameterResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/processes/{processId}/parameters/{id}": {
      "get": {
        "tags": [
          "Process Parameter Master"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProcessParameterResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Process Parameter Master"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Roles: SUPER_ADMIN, ADMIN",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProcessParameterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProcessParameterResponse"
                    }
                  }
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Process Parameter Master"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Roles: SUPER_ADMIN",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ParameterSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateProcessParameterRequest": {
        "type": "object",
        "properties": {
          "parameterId": {
            "type": "integer",
            "format": "int64"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "mandatory": {
            "type": "boolean"
          },
          "visible": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "minimumValue": {
            "type": "number"
          },
          "maximumValue": {
            "type": "number"
          }
        }
      },
...
      "ProcessParameterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "processId": {
            "type": "integer",
            "format": "int64"
          },
          "parameter": {
            "$ref": "#/components/schemas/ParameterSummaryResponse"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "mandatory": {
            "type": "boolean"
          },
          "visible": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "minimumValue": {
            "type": "number"
          },
          "maximumValue": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "inputType": {
            "type": "string",
            "enum": [
              "NUMBER",
              "// Numeric input\n    TEXT",
              "// Text input\n    BOOLEAN",
              "// Boolean toggle input\n    DROPDOWN // Dropdown selection input"
            ]
          }
        }
      },
...
      "UpdateProcessParameterRequest": {
        "type": "object",
        "properties": {
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "mandatory": {
            "type": "boolean"
          },
          "visible": {
            "type": "boolean"
          },
          "defaultValue": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "minimumValue": {
            "type": "number"
          },
          "maximumValue": {
            "type": "number"
          }
        }
      },
...
    }
  }
}